backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
google-cloud-compute==1.14.1
google-cloud-monitoring==2.16.0
google-cloud-pubsub==2.18.4
google-cloud-spanner==3.40.1
kubernetes==28.1.0
httpx==0.25.1
pydantic==2.5.0
python-dotenv==1.0.0
PyYAML==6.0.1
google-auth==2.23.4
